Ajustement de la page /consulter pour supporter un /entityFilter et activer le bouton de filtre sur le fetch de la page
